(TK-316) Remove subclasses from ClientMetricFilter
Previously all the filter options were implemented as subclasses of the
ClientMetricFilter class. This commit changes it so that ClientMetricFilter
has two constructors, one that takes a string and one that takes no arguments,
and all the logic of constructing the name to match against is moved to the
Metric class where it is used.